+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Release `1.5.1` - *IN PROGRESS*
 
 - Fix some variables that cannot be set via api
+- New action on key events to set a variable
 
 
 ## Release `1.5` - *2021-06-23*